HOTT-4523 Refactored ETag caching
