+++ app/controllers/api/v2/exchange_rates/base_controller.rb
@@ -2,7 +2,7 @@ module Api
   module V2
     module ExchangeRates
       class BaseController < ApiController
-        include NoCaching
+        no_caching
 
         before_action :validate_exchange_rate_type
