+++ app/controllers/api/v2/news/items_controller.rb
@@ -2,7 +2,7 @@ module Api
   module V2
     module News
       class ItemsController < ApiController
-        include SimpleCaching
+        time_based_caching
 
         PAGE_SIZES = [1, 10, 20].freeze
         DEFAULT_PAGE_SIZE = 20
