--- app/controllers/concerns/no_caching.rb
@@ -1,7 +1,0 @@
-module NoCaching
-  protected
-
-  def set_cache_headers
-    no_store
-  end
-end
